+++ mfkdf2/src/lib.rs
@@ -1,7 +1,7 @@
 uniffi::setup_scaffolding!();
 
-pub mod classes;
 pub mod crypto;
+pub mod definitions;
 pub mod derive;
 pub mod error;
 pub mod integrity;
