+++ mfkdf2/src/policy/setup.rs
@@ -1,5 +1,5 @@
 use crate::{
-  classes::mfkdf_derived_key::MFKDF2DerivedKey,
+  definitions::mfkdf_derived_key::MFKDF2DerivedKey,
   error::{MFKDF2Error, MFKDF2Result},
   setup::{
     factors::MFKDF2Factor,
